+++ astroquery/gaia/tests/test_gaia_remote.py
@@ -56,3 +56,85 @@ def test_cone_search_row_limit():
     r = j.get_results()
 
     assert len(r) == 1218
+
+@pytest.mark.remote_data
+def test_query_async_object_columns_with_epoch_prop():
+    # Regression test: `columns` were ignored if `radius` was provided [#2025]
+    gaia = GaiaClass()
+    w = u.Quantity(0.001, u.deg)
+    table = gaia.query_object_async('Proxima Centauri', width=w, height=w)
+    assert table['source_id'] == '5853498678802473344'
+
+    table = gaia.query_object_async('Proxima Centauri', width=w, height=w, epoch_prop=True)
+    assert table['source_id'] == '5853498713160606720'
+
+
+@pytest.mark.remote_data
+def test_query_sync_object_columns_with_epoch_prop():
+    # Regression test: `columns` were ignored if `radius` was provided [#2025]
+    gaia = GaiaClass()
+    w = u.Quantity(0.001, u.deg)
+    table = gaia.query_object('Proxima Centauri', width=w, height=w)
+    assert table['source_id'] == '5853498678802473344'
+
+    table = gaia.query_object('Proxima Centauri', width=w, height=w, epoch_prop=True)
+    assert table['source_id'] == '5853498713160606720'
+
+def test_cone_search_async_epoch_prop():
+    gaia = GaiaClass()
+    gaia.ROW_LIMIT = -1
+
+    #ICRS coord. (ep=J2000) : 14:29:42.9461331854 -62:40:46.164680672
+    coord = SkyCoord(ra='14:29:42.9461331854', dec='-62:40:46.164680672', unit=(u.hourangle, u.deg), frame='icrs')
+    radius = u.Quantity(0.1, u.deg)
+
+    job = gaia.cone_search_async(coord, radius=radius)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498678802473344'
+
+    job = gaia.cone_search_async(coord, radius=radius, epoch_prop=True)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498713160606720'
+
+
+    job = gaia.cone_search_async('Proxima Centauri', radius=radius)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498678802473344'
+
+    job = gaia.cone_search_async('Proxima Centauri', radius=radius, epoch_prop=True)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498713160606720'
+
+def test_cone_search_epoch_prop():
+    gaia = GaiaClass()
+    gaia.ROW_LIMIT = -1
+
+    #ICRS coord. (ep=J2000) : 14:29:42.9461331854 -62:40:46.164680672
+    coord = SkyCoord(ra='14:29:42.9461331854', dec='-62:40:46.164680672', unit=(u.hourangle, u.deg), frame='icrs')
+    radius = u.Quantity(0.1, u.deg)
+
+    job = gaia.cone_search(coord, radius=radius)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498678802473344'
+
+    job = gaia.cone_search(coord, radius=radius, epoch_prop=True)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498713160606720'
+
+
+    job = gaia.cone_search('Proxima Centauri', radius=radius)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498678802473344'
+
+    job = gaia.cone_search('Proxima Centauri', radius=radius, epoch_prop=True)
+    table = job.get_results()
+    assert len(table) == 1218
+    assert table['source_id'] == '5853498713160606720'
+